Move PagesMenu to enclosing folder
